Release - v1.0
With this release, the script is ready for use. Bugs have been fixed from the beta version and presentation of output to the user has been brought to an acceptable level (acceptable being: easier to read than the default Caliber CV). Development will remain ongoing depending on the performance of this script, and improvements will be made in later versions.

Patch notes (v1.0):
- Fixed bug that prevented non-highlighted words that occurred after the keyword to be printed
- Sentences are now dislayed as bullet points with added indent on all lines but the first.
- Text is now wrapped so words are not split up if they reach end-of-line.
- Newlines added and deleted in certain places for tidier output.
- Adjusted code to fit window size of width=70
- Broad skill headings now coloured yellow to indicate they're separate from specific skill headings.